Added long to Alarm Receiver.

+++ app/src/main/java/com/example/trickyalarm/AlarmReceiver.java
@@ -33,7 +33,7 @@ public class AlarmReceiver extends BroadcastReceiver {
      * @param context
      * @param time in which alarm should work.
      */
-    public void setAlarm(Context context, int time) {
+    public void setAlarm(Context context, long time) {
         AlarmManager alarmManager = (AlarmManager) context.getSystemService(Context.ALARM_SERVICE);
         Intent intent = new Intent(context, AlarmReceiver.class);
         PendingIntent pendingIntent = PendingIntent.getBroadcast(context, 0, intent, 0);
